refactor: split the SDDL parser out of windows-acl.ts
`parseDacl` is the security decision for the whole Windows config path, and the
reason its 2000-run property tests can guard that decision on the Linux job is
that the parser touches nothing platform-specific. That was true only by
discipline: an edit reaching for `run(ICACLS, …)` inside `parseDacl` would
compile, read as a one-liner, and quietly turn every parser test into a
Windows-only test — while the suite stayed green, because the property and unit
files would simply stop exercising it off Windows.

src/config/sddl.ts is the grammar half: the SID and alias tables, the rights
masks, `grantsWrite`, the ACE scanners (`quoteBlindBalanced`, `aceGroups`,
`daclFlagsAreValid`, `aceFlagCodes`), `resolveTrustee`, `parseDacl`,
`aclIdentity`, and the `Grant`/`UnknownReason`/`AclVerdict`/`AclIdentity` types.
It imports nothing at all. windows-acl.ts keeps everything that talks to the
machine — `icacls`, the current SID, the remediation text, the refuse-or-report
decision — and drops from 1055 to 595 lines.

A separate file does not create the guarantee on its own, so
sddl-import-boundary.test.ts asserts it: sddl.ts must have an empty import list,
must not name a platform module by any spelling (`require`, dynamic `import()`,
or a side-effect-only `import 'os'` — the last slips past a `from`-anchored
scan), and must load without pulling one in transitively. Verified: all three
sneak-paths fail the test.

Two deviations from the queued plan, both because the code disagreed with it:

  - `PRINCIPAL_NAMES`/`nameFor` were queued as belonging on the message side.
    They do not: `parseDacl` calls `nameFor` to populate `Grant.name`, so moving
    them out would have meant giving the parser a name-resolver parameter, which
    is not a pure move. They travel with the parser.
  - `classifyReadFailure` stays. It is pure, but it interprets an `icacls`
    failure, which is the platform half's vocabulary.

No export markers had to change — the existing ones already sat exactly on the
seam, which is the clearest sign the cut is where the module already wanted it.

Pure move, verified as one: 467 relocated lines, all present verbatim in the new
file, and no distinctive line left duplicated behind. Three test files repoint;
two now import from both sides, which is the dependency the split exists to make
visible. No changeset — the package publishes only build/ behind a bin.

Verified: 674 tests against the live containers, 29 e2e.
